last comit

+++ Accompany-consulting/Accompany-consulting/Models/Consultant.cs
@@ -54,7 +54,7 @@ namespace Accompany_consulting.Models
         public string code { get; set; }
         public string situation_amoureuse { get; set; }
         public int salaire { get; set; }
-        public int SoldeConge { get; set; }
-        public int SoldeMaladie { get; set; }
+        public double SoldeConge { get; set; }
+        public double SoldeMaladie { get; set; }
     }
 }
